+++ lldb/source/Plugins/SymbolLocator/Debuginfod/CMakeLists.txt
@@ -14,6 +14,7 @@ add_lldb_library(lldbPluginSymbolLocatorDebuginfod PLUGIN
     lldbHost
     lldbSymbol
     LLVMDebuginfod
+    LLVMSupportHTTP
   )
 
 add_dependencies(lldbPluginSymbolLocatorDebuginfod
