+++ llvm/lib/Support/CMakeLists.txt
@@ -135,6 +135,7 @@ if (UNIX AND ${CMAKE_SYSTEM_NAME} MATCHES "AIX")
 endif()
 
 add_subdirectory(BLAKE3)
+add_subdirectory(HTTP)
 
 add_llvm_component_library(LLVMSupport
   ABIBreak.cpp
