+++ llvm/tools/llvm-cov/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(LLVM_LINK_COMPONENTS
   Core
   Support
+  SupportHTTP
   Object
   Coverage
   ProfileData
